requirements.txt
# Core
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0

# ML
scikit-learn>=1.1.0
torch>=2.0.0
transformers>=4.30.0

# Bio
biopython>=1.80
obonet>=0.3.0
networkx>=2.8

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.15.0

# Utils
tqdm>=4.64.0

# Config / notebook tooling
pyyaml>=6.0
nbformat>=5.9.0
nbclient>=0.8.0
